+++ packages/evokit-body/__tests__/Body.display.test.jsx
@@ -1,5 +1,7 @@
 import React from 'react';
 import { shallow } from 'enzyme';
+import barista from 'seed-barista';
+import style from '../style.css';
 import { Body } from '../src';
 
 describe('<Body body-display="..." />', () => {
@@ -8,5 +10,12 @@ describe('<Body body-display="..." />', () => {
         expect(wrapper).toMatchSnapshot();
 
         expect(wrapper.html()).toEqual('<body class="ek-body ek-body_display_value"></body>');
+    });
+
+    it('css', () => {
+        const css = barista({ content: style });
+
+        expect(css.rule('.ek-body_display_block').exists()).toBeTruthy();
+        expect(css.rule('.ek-body_display_none').exists()).toBeTruthy();
     });
 });